plugins/stashStashboxSceneCount/requirements.txt
requests==2.30.0
pystashlib==0.4.4